Moved gestures into separated files
