+++ Examples/Nodify.Shapes/Canvas/CanvasToolbarViewModel.cs
@@ -16,7 +16,7 @@ namespace Nodify.Shapes.Canvas
     {
         public static readonly CanvasTool[] AvailableTools = Enum.GetValues(typeof(CanvasTool)).Cast<CanvasTool>().ToArray();
 
-        internal static readonly EditorGestures EditorGestures;
+        internal static readonly EditorGestures EditorGestures = new EditorGestures();
 
         private bool _locked;
         public bool Locked
@@ -56,52 +56,13 @@ namespace Nodify.Shapes.Canvas
 
         public CanvasViewModel Canvas { get; }
 
-        static CanvasToolbarViewModel()
-        {
-            EditorGestures = new EditorGestures();
-            // copy any user modifications
-            EditorGestures.Apply(EditorGestures.Mappings);
-        }
-
         public CanvasToolbarViewModel(CanvasViewModel canvas)
         {
+            // copy any user modifications
+            EditorGestures.Apply(EditorGestures.Mappings);
+
             ToggleLockCommand = new DelegateCommand(() => Locked = !Locked);
             Canvas = canvas;
-        }
-    }
-
-    public class UnboundGestureMappings : EditorGestures
-    {
-        public static readonly UnboundGestureMappings Instance = new UnboundGestureMappings();
-
-        public UnboundGestureMappings()
-        {
-            Editor.Selection.Apply(SelectionGestures.None);
-            ItemContainer.Selection.Apply(SelectionGestures.None);
-            Connection.Disconnect.Value = MultiGesture.None;
-            Connector.Connect.Value = MultiGesture.None;
-        }
-    }
-
-    public class LockedGestureMappings : EditorGestures
-    {
-        public static readonly LockedGestureMappings Instance = new LockedGestureMappings();
-
-        public LockedGestureMappings()
-        {
-            Apply(UnboundGestureMappings.Instance);
-
-            Editor.Pan.Value = new AnyGesture(new MouseGesture(MouseAction.LeftClick), new MouseGesture(MouseAction.RightClick), new MouseGesture(MouseAction.MiddleClick));
-        }
-    }
-
-    public class DrawingGesturesMappings : EditorGestures
-    {
-        public static readonly DrawingGesturesMappings Instance = new DrawingGesturesMappings();
-
-        public DrawingGesturesMappings()
-        {
-            Apply(UnboundGestureMappings.Instance);
         }
     }
 }
