+++ Examples/Nodify.Shapes/Canvas/Gestures/DrawingGesturesMappings.cs
@@ -1,0 +1,18 @@
+﻿using System.Windows.Input;
+
+namespace Nodify.Shapes.Canvas
+{
+    public class DrawingGesturesMappings : EditorGestures
+    {
+        public static readonly DrawingGesturesMappings Instance = new DrawingGesturesMappings();
+
+        public InputGestureRef Draw { get; }
+
+        public DrawingGesturesMappings()
+        {
+            Apply(UnboundGestureMappings.Instance);
+
+            Draw = new MouseGesture(MouseAction.LeftClick);
+        }
+    }
+}
